// lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FinTrack API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  ingreso: "ingreso",
  gasto: "gasto",
} as const;

export interface Transaction {
  id: number;
  name: string;
  cat: string;
  amount: number;
  type: TransactionType;
  date: string;
  accountId?: number | null;
  createdAt: string;
}

export type CreateTransactionInputType =
  (typeof CreateTransactionInputType)[keyof typeof CreateTransactionInputType];

export const CreateTransactionInputType = {
  ingreso: "ingreso",
  gasto: "gasto",
} as const;

export interface CreateTransactionInput {
  name: string;
  cat: string;
  amount: number;
  type: CreateTransactionInputType;
  date: string;
  accountId?: number | null;
}

export interface Budget {
  id: number;
  cat: string;
  limit: number;
  color: string;
}

export interface CreateBudgetInput {
  cat: string;
  limit: number;
  color?: string;
}

export interface Account {
  id: number;
  name: string;
  bank: string;
  bal: number;
  color: string;
}

export interface CreateAccountInput {
  name: string;
  bank?: string;
  bal: number;
}

export interface Goal {
  id: number;
  name: string;
  saved: number;
  target: number;
  color: string;
}

export interface CreateGoalInput {
  name: string;
  saved?: number;
  target: number;
  color?: string;
}

export interface RecurringPayment {
  id: number;
  name: string;
  amount: number;
  dayOfMonth: number;
  accountId?: number | null;
}

export interface CreateRecurringInput {
  name: string;
  amount: number;
  dayOfMonth: number;
  accountId?: number | null;
}

export interface Category {
  id: number;
  name: string;
  icon?: string | null;
  color?: string | null;
}